Refactor some code
and add documentation
